requirements.txt
asgiref==3.7.2
cloudinary==1.38.0
Django==3.2.23
django-cloudinary-storage==0.3.0
django-filter==23.5
djangorestframework==3.14.0
pillow==10.2.0
pytz==2023.3.post1
sqlparse==0.4.4
